refactor: move resolve_tags to shared utility and wire completed/due_date into create_todo (#nirvana-90)
- Extract resolve_tags (and _infer_tag_type) from todos/routes.py into
  new todos/utils.py so it can be imported as a public symbol by other
  modules, removing the fragile _resolve_tags private import
- Update both todos/routes.py and import_nirvana/routes.py to import
  resolve_tags from todos.utils
- Wire completed and due_date into the create_todo handler so values
  passed to POST /todos are not silently dropped
- Update test import to match new module location
